add check for file exists error

+++ command-injection/apex-publish-static-files_2.0.0/apex-publish-static-files.test.js
@@ -5,8 +5,10 @@ test("Command Injection in apex-publish-static-files", () => {
   const path = "./apex-publish-static-files";
   const sleep = require("sleep");
   try {
-    fs.unlinkSync(path);
-    console.log("File removed:", path);
+    if (fs.existsSync(path)) {
+      fs.unlinkSync(path);
+      console.log("File removed:", path);
+    }
   } catch (err) {
     console.error(err);
   }
