+++ command-injection/effect_1.0.4/effect.test.js
@@ -6,8 +6,10 @@ test("Command Injection in effect", () => {
   const sleep = require("sleep");
   var options = { image: "& touch effect" };
   try {
-    fs.unlinkSync(path);
-    console.log("File removed:", path);
+    if (fs.existsSync(path)) {
+      fs.unlinkSync(path);
+      console.log("File removed:", path);
+    }
   } catch (err) {
     console.error(err);
   }
