+++ command-injection/ps-kill_1.0.0/ps-kill.test.js
@@ -5,8 +5,10 @@ test("Command Injection in ps-kill", () => {
   const path = "./ps-kill";
   const sleep = require("sleep");
   try {
-    fs.unlinkSync(path);
-    console.log("File removed:", path);
+    if (fs.existsSync(path)) {
+      fs.unlinkSync(path);
+      console.log("File removed:", path);
+    }
   } catch (err) {
     console.error(err);
   }
