+++ command-injection/samsung-remote_1.2.5/samsung-remote.test.js
@@ -5,8 +5,10 @@ test("Command Injection in samsung-remote", () => {
   const path = "./samsung-remote";
   const sleep = require("sleep");
   try {
-    fs.unlinkSync(path);
-    console.log("File removed:", path);
+    if (fs.existsSync(path)) {
+      fs.unlinkSync(path);
+      console.log("File removed:", path);
+    }
   } catch (err) {
     console.error(err);
   }
