+++ command-injection/ts-process-promises_1.0.2/ts-process-promises.test.js
@@ -5,8 +5,10 @@ test("Command Injection in ts-process-promises", () => {
   const path = "./ts-process-promises";
   const sleep = require("sleep");
   try {
-    fs.unlinkSync(path);
-    console.log("File removed:", path);
+    if (fs.existsSync(path)) {
+      fs.unlinkSync(path);
+      console.log("File removed:", path);
+    }
   } catch (err) {
     console.error(err);
   }
